Split worktree info monitor implementations
